feat: rename catalog::local to catalog::memory
