+++ src/operator/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "operator"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[features]
+testing = []
+
+[dependencies]
+
+[dev-dependencies]
