+++ src/query/src/tests/query_engine_test.rs
@@ -14,7 +14,7 @@
 
 use std::sync::Arc;
 
-use catalog::local::MemoryCatalogManager;
+use catalog::memory::MemoryCatalogManager;
 use catalog::RegisterTableRequest;
 use common_base::Plugins;
 use common_catalog::consts::{DEFAULT_CATALOG_NAME, DEFAULT_SCHEMA_NAME, NUMBERS_TABLE_ID};
@@ -44,7 +44,7 @@ use crate::tests::pow::pow;
 #[tokio::test]
 async fn test_datafusion_query_engine() -> Result<()> {
     common_telemetry::init_default_ut_logging();
-    let catalog_list = catalog::local::new_memory_catalog_manager()
+    let catalog_list = catalog::memory::new_memory_catalog_manager()
         .map_err(BoxedError::new)
         .context(QueryExecutionSnafu)?;
     let factory = QueryEngineFactory::new(catalog_list, None, false);
@@ -104,7 +104,7 @@ async fn test_datafusion_query_engine() -> Result<()> {
 }
 
 fn catalog_manager() -> Result<Arc<MemoryCatalogManager>> {
-    let catalog_manager = catalog::local::new_memory_catalog_manager().unwrap();
+    let catalog_manager = catalog::memory::new_memory_catalog_manager().unwrap();
     let req = RegisterTableRequest {
         catalog: DEFAULT_CATALOG_NAME.to_string(),
         schema: DEFAULT_SCHEMA_NAME.to_string(),
